+++ src/FluxFlow.Components.Configuration/Contracts/ConfigurationDiagnosticSeverity.cs
@@ -1,0 +1,8 @@
+namespace FluxFlow.Components.Configuration.Contracts;
+
+public enum ConfigurationDiagnosticSeverity
+{
+    Information = 1,
+    Warning = 2,
+    Error = 3
+}